Changes requested in Pull Request #173
